requirements.txt
discord.py
dblpy
